% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.match_type.r
\name{as.match_type}
\alias{as.match_type}
\title{As Match Type}
\usage{
as.match_type(valuetrak_match_type_vector)
}
\arguments{
\item{valuetrak_match_type_vector}{A vector of match type valuetrak values.}
}
\description{
This function converts the non-human-readable values in the matchtype
vector provided by Google AdWords valuetrak tags to a human-readable
version.
}
\examples{
as.match_type(valuetrak_match_type_vector)
as.match_type(google_adwords_report$match_type)
}

