% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_adwords_elog.r
\name{summarize_adwords_elog}
\alias{summarize_adwords_elog}
\title{Summarize AdWords Event Log}
\usage{
summarize_adwords_elog(elog_data_frame, margin = 0.25,
  average_num_orders_in_lifetime = 10, average_value_per_order = 70)
}
\arguments{
\item{elog_data_frame}{A data frame containing all of the AdWords Event Log data.}

\item{margin}{Estimated percentage of contribution margin (e.g. 0.25 for 25%). Used in contribution and estimated Return on Ad Spend (ROAS) calculations}

\item{average_num_orders_in_lifetime}{The number of orders a new customer makes, on average. Used in estimated Return on Ad Spend (ROAS) calculations}

\item{average_value_per_order}{The average value per order.  Used in estimated Return on Ad Spend (ROAS) calculations}
}
\description{
Takes in an AdWords event log and processes all summary data
}
\examples{
summarize_adwords_elog(elog_data_frame)
}

