% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.adwords.csv.r
\name{write.excel.csv}
\alias{write.excel.csv}
\title{Write Excel CSV}
\usage{
write.excel.csv(data_frame, file)
}
\arguments{
\item{data_frame}{Data frame to be written to CSV file.}

\item{file}{The path to the file to be written}
}
\description{
Wrapper function to write data frame to an excel-readable CSV file.
}
\examples{
write.excel.csv(data_frame, "some_csv_file.csv")
}

